improving http requests

+++ apirer/src/main/java/com/apirergr/apirer/controllers/VendedoresController.java
@@ -1,13 +1,17 @@
 package com.apirergr.apirer.controllers;
+
 import com.apirergr.apirer.modelos.Vendedores;
 import com.apirergr.apirer.repositorios.VendedoresRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/vendedores")
 public class VendedoresController {
+
     @Autowired
     private VendedoresRepo vendedoresRepo;
 
@@ -17,8 +21,10 @@ public class VendedoresController {
     }
 
     @GetMapping("/{id}")
-    public Vendedores obtenerVendedorPorId(@PathVariable Long id) {
-        return vendedoresRepo.findById(id).orElse(null);
+    public ResponseEntity<Vendedores> obtenerVendedorPorId(@PathVariable Long id) {
+        Optional<Vendedores> vendedor = vendedoresRepo.findById(id);
+        return vendedor.map(ResponseEntity::ok)
+                .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @PostMapping
@@ -27,18 +33,25 @@ public class VendedoresController {
     }
 
     @PutMapping("/{id}")
-    public Vendedores actualizarVendedor(@PathVariable Long id, @RequestBody Vendedores vendedorActualizado) {
+    public ResponseEntity<Vendedores> actualizarVendedor(@PathVariable Long id, @RequestBody Vendedores vendedorActualizado) {
         return vendedoresRepo.findById(id)
                 .map(vendedor -> {
                     vendedor.setNombre(vendedorActualizado.getNombre());
-                    return vendedoresRepo.save(vendedor);
+                    vendedor.setDNI(vendedorActualizado.getDNI());
+                    vendedor.setApellido(vendedorActualizado.getApellido());
+                    vendedor.setExperiencia(vendedorActualizado.getExperiencia());
+                    return ResponseEntity.ok(vendedoresRepo.save(vendedor));
                 })
-                .orElse(null);
+                .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
-    public void eliminarVendedor(@PathVariable Long id) {
-        vendedoresRepo.deleteById(id);
+    public ResponseEntity<Void> eliminarVendedor(@PathVariable Long id) {
+        if (vendedoresRepo.existsById(id)) {
+            vendedoresRepo.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
-
 }
